requirements.txt
gradio
faster-whisper
langchain
langgraph
langchain-openai
openai
notion-client
python-dotenv
pyyaml
dateparser
langfuse
